extract navbar and let it changed depending if user is logged in
