+++ app/layout.tsx
@@ -3,6 +3,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import Link from "next/link";
 import { ClerkProvider, UserButton } from "@clerk/nextjs";
+import NavBar from "./components/NavBar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -10,14 +11,6 @@ export const metadata: Metadata = {
   title: "Create Next App",
   description: "Generated by create next app",
 };
-
-const links = [
-  { href: "/", label: "Home" },
-  { href: "/online-order-refund", label: "Online order in-store refund" },
-  { href: "/server-form", label: "Server Form" },
-  { href: "/sign-in", label: "Sign in" },
-  { href: "/sign-up", label: "Sign up" },
-];
 
 export default function RootLayout({
   children,
@@ -29,18 +22,7 @@ export default function RootLayout({
       <html lang="en">
         <body className={inter.className}>
           <header>
-            <nav>
-              <ul className="flex items-center">
-                {links.map((link) => (
-                  <li key={link.href}>
-                    <Link href={link.href}>{link.label}</Link>
-                  </li>
-                ))}
-                <li>
-                  <UserButton afterSignOutUrl="/" />
-                </li>
-              </ul>
-            </nav>
+            <NavBar />
           </header>
 
           <div>{children}</div>
